+++ lib/ui/src/containers/nav.test.js
@@ -1,4 +1,4 @@
-import { collapseDocsOnlyStories, collapseAllStories } from './nav';
+import { collapseDocsOnlyStories, collapseAllStories } from './sidebar';
 
 const docsOnly = { parameters: { docsOnly: true } };
 const root = { id: 'root', parent: false, children: ['a', 'b'] };
